+++ tests/redis_transports/test_integration_redis.py
@@ -15,7 +15,7 @@ from lightbus.transports.redis.event import StreamUse
 from lightbus.utilities.async_tools import cancel
 from lightbus.utilities.features import Feature
 from lightbus.utilities.testing import BusQueueMockerContext
-from tests.redis_transports.conftest import Worker
+from tests.conftest import Worker
 
 pytestmark = pytest.mark.integration