+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 0.10.0 - Unreleased
 
-* ...
+* The `on_error` configuration parameter has been removed. All errors are now fatal to the 
+  Lightbus worker. If you wish for errors to not be fatal you should catch them in your 
+  event handlers.
 
 ## 0.9.1 - Unreleased